+++ src/main/java/com/cloudbees/jenkins/plugins/amazonecs/ECSLauncher.java
@@ -25,7 +25,9 @@
 
 package com.cloudbees.jenkins.plugins.amazonecs;
 
-import static java.util.logging.Level.*;
+import static java.util.logging.Level.FINE;
+import static java.util.logging.Level.INFO;
+import static java.util.logging.Level.SEVERE;
 
 import java.io.IOException;
 import java.io.PrintStream;
@@ -36,16 +38,14 @@ import java.util.Collection;
 import java.util.logging.Level;
 import java.util.logging.Logger;
 
-import org.apache.commons.lang.StringUtils;
-import org.kohsuke.stapler.DataBoundConstructor;
-
-import com.amazonaws.services.ecs.model.DescribeTasksRequest;
-import com.amazonaws.services.ecs.model.DescribeTasksResult;
 import com.amazonaws.services.ecs.model.Failure;
 import com.amazonaws.services.ecs.model.RunTaskResult;
 import com.amazonaws.services.ecs.model.Task;
 import com.amazonaws.services.ecs.model.TaskDefinition;
 import com.google.common.base.Throwables;
+
+import org.apache.commons.lang.StringUtils;
+import org.kohsuke.stapler.DataBoundConstructor;
 
 import hudson.AbortException;
 import hudson.model.TaskListener;
@@ -89,28 +89,28 @@ public class ECSLauncher extends JNLPLauncher {
         ECSComputer ecsComputer = (ECSComputer) computer;
         computer.setAcceptingTasks(false);
 
-        ECSSlave slave = (ECSSlave)ecsComputer.getNode();
-        if (slave == null) {
+        ECSSlave agent = (ECSSlave)ecsComputer.getNode();
+        if (agent == null) {
             throw new IllegalStateException("Node has been removed, cannot launch " + computer.getName());
         }
 
         if (launched) {
-            LOGGER.log(INFO, "[{0}]: Agent has already been launched, activating", slave.getNodeName());
+            LOGGER.log(INFO, "[{0}]: Agent has already been launched, activating", agent.getNodeName());
             computer.setAcceptingTasks(true);
             return;
         }
 
         try {
-            LOGGER.log(Level.FINE, "[{0}]: Creating Task in cluster {1}", new Object[]{slave.getNodeName(), slave.getClusterArn()});
+            LOGGER.log(Level.FINE, "[{0}]: Creating Task in cluster {1}", new Object[]{agent.getNodeName(), agent.getClusterArn()});
 
-            TaskDefinition taskDefinition = getTaskDefinition(slave.getNodeName(), slave.getTemplate(), cloud, ecsService);
+            TaskDefinition taskDefinition = getTaskDefinition(agent.getNodeName(), agent.getTemplate(), cloud, ecsService);
 
-            Task task = runECSTask(taskDefinition, cloud, slave.getTemplate(), ecsService, slave);
+            Task task = runECSTask(taskDefinition, cloud, agent.getTemplate(), ecsService, agent);
 
-            LOGGER.log(INFO, "[{0}]: TaskArn: {1}", new Object[]{slave.getNodeName(), task.getTaskArn()});
-            LOGGER.log(INFO, "[{0}]: TaskDefinitionArn: {1}", new Object[]{slave.getNodeName(), task.getTaskDefinitionArn()});
-            LOGGER.log(INFO, "[{0}]: ClusterArn: {1}", new Object[]{slave.getNodeName(), task.getClusterArn()});
-            LOGGER.log(INFO, "[{0}]: ContainerInstanceArn: {1}", new Object[]{slave.getNodeName(), task.getContainerInstanceArn()});
+            LOGGER.log(INFO, "[{0}]: TaskArn: {1}", new Object[]{agent.getNodeName(), task.getTaskArn()});
+            LOGGER.log(INFO, "[{0}]: TaskDefinitionArn: {1}", new Object[]{agent.getNodeName(), task.getTaskDefinitionArn()});
+            LOGGER.log(INFO, "[{0}]: ClusterArn: {1}", new Object[]{agent.getNodeName(), task.getClusterArn()});
+            LOGGER.log(INFO, "[{0}]: ContainerInstanceArn: {1}", new Object[]{agent.getNodeName(), task.getContainerInstanceArn()});
 
             long timeout = System.currentTimeMillis() + Duration.ofSeconds(cloud.getSlaveTimoutInSeconds()).toMillis();
 
@@ -122,59 +122,59 @@ public class ECSLauncher extends JNLPLauncher {
 
                 if (task != null) {
                     String taskStatus = task.getLastStatus();
-                    LOGGER.log(Level.FINE, "[{0}]: Status of ECS Task is {1}", new Object[]{slave.getNodeName(), taskStatus});
+                    LOGGER.log(Level.FINE, "[{0}]: Status of ECS Task is {1}", new Object[]{agent.getNodeName(), taskStatus});
 
                     if (taskStatus.equals("RUNNING")) {
                         taskRunning = true;
                         break;
                     }
                     if (taskStatus.equals("STOPPED")) {
-                        LOGGER.log(Level.WARNING, "[{0}]: ECS Task stopped: {1}", new Object[]{slave.getNodeName(), task.getStoppedReason()});
+                        LOGGER.log(Level.WARNING, "[{0}]: ECS Task stopped: {1}", new Object[]{agent.getNodeName(), task.getStoppedReason()});
                         logger.printf("ECS Task stopped: %1$s%n", task.getStoppedReason());
                         throw new IllegalStateException("Task stopped before coming online");
                     }
                 }
 
-                LOGGER.log(INFO, "[{0}]: Waiting for agent to start", new Object[]{slave.getNodeName()});
-                logger.printf("Waiting for agent to start: %1$s%n", slave.getNodeName());
+                LOGGER.log(INFO, "[{0}]: Waiting for agent to start", new Object[]{agent.getNodeName()});
+                logger.printf("Waiting for agent to start: %1$s%n", agent.getNodeName());
                 Thread.sleep(1000);
             }
 
             if (!taskRunning) {
                 if (task != null) {
-                    LOGGER.log(SEVERE, "[{0}]: Task is not running. Last status: {1}, Exit code: {2}, Reason {3}", new Object[]{slave.getNodeName(), task.getLastStatus(), task.getContainers().get(0).getExitCode(), task.getContainers().get(0).getReason()});
+                    LOGGER.log(SEVERE, "[{0}]: Task is not running. Last status: {1}, Exit code: {2}, Reason {3}", new Object[]{agent.getNodeName(), task.getLastStatus(), task.getContainers().get(0).getExitCode(), task.getContainers().get(0).getReason()});
                 }
                 throw new IllegalStateException("Task took too long to start");
             }
 
-            LOGGER.log(INFO, "[{0}]: Task started, waiting for agent to become online", new Object[]{slave.getNodeName()});
+            LOGGER.log(INFO, "[{0}]: Task started, waiting for agent to become online", new Object[]{agent.getNodeName()});
 
             // now wait for agent to be online
             while (System.currentTimeMillis() < timeout) {
 
-                if (slave.getComputer() == null) {
+                if (agent.getComputer() == null) {
                     throw new IllegalStateException("Node was deleted, computer is null");
                 }
-                if (slave.getComputer().isOnline()) {
+                if (agent.getComputer().isOnline()) {
                     break;
                 }
-                LOGGER.log(INFO, "[{0}]: Waiting for agent to connect", new Object[]{slave.getNodeName()});
-                logger.printf("Waiting for agent to connect: %1$s%n", slave.getNodeName());
+                LOGGER.log(INFO, "[{0}]: Waiting for agent to connect", new Object[]{agent.getNodeName()});
+                logger.printf("Waiting for agent to connect: %1$s%n", agent.getNodeName());
                 Thread.sleep(1000);
             }
 
-            if (!slave.getComputer().isOnline()) {
+            if (!agent.getComputer().isOnline()) {
                 throw new IllegalStateException("Agent is not connected");
             }
-            LOGGER.log(INFO, "[{0}]: Agent connected", new Object[]{slave.getNodeName()});
+            LOGGER.log(INFO, "[{0}]: Agent connected", new Object[]{agent.getNodeName()});
 
             computer.setAcceptingTasks(true);
 
         } catch (Throwable ex) {
-            LOGGER.log(Level.WARNING, MessageFormat.format("[{0}]: Error in provisioning; agent={1}", slave.getNodeName(), slave), ex);
-            LOGGER.log(Level.FINER, "[{0}]: Removing Jenkins node", slave.getNodeName());
+            LOGGER.log(Level.WARNING, MessageFormat.format("[{0}]: Error in provisioning; agent={1}", agent.getNodeName(), agent), ex);
+            LOGGER.log(Level.FINER, "[{0}]: Removing Jenkins node", agent.getNodeName());
             try {
-                slave.terminate();
+                agent.terminate();
             } catch (InterruptedException e) {
                 LOGGER.log(Level.WARNING, "Unable to remove Jenkins node", e);
             } catch (IOException e) {
@@ -187,7 +187,7 @@ public class ECSLauncher extends JNLPLauncher {
 
         try {
             // We need to persist the "launched" setting...
-            slave.save();
+            agent.save();
         } catch (IOException e) {
             LOGGER.log(Level.WARNING, "Could not save() agent: " + e.getMessage(), e);
         }
@@ -213,26 +213,25 @@ public class ECSLauncher extends JNLPLauncher {
         return taskDefinition;
     }
 
-    private Task runECSTask(TaskDefinition taskDefinition, ECSCloud cloud, ECSTaskTemplate template, ECSService ecsService, ECSSlave slave) throws IOException, AbortException {
+    private Task runECSTask(TaskDefinition taskDefinition, ECSCloud cloud, ECSTaskTemplate template, ECSService ecsService, ECSSlave agent) throws IOException, AbortException {
 
+        LOGGER.log(Level.INFO, "[{0}]: Starting agent with task definition {1}}", new Object[]{agent.getNodeName(), taskDefinition.getTaskDefinitionArn()});
 
-        LOGGER.log(Level.INFO, "[{0}]: Starting agent with task definition {1}}", new Object[]{slave.getNodeName(), taskDefinition.getTaskDefinitionArn()});
-
-        RunTaskResult runTaskResult = ecsService.runEcsTask(slave, template, cloud.getCluster(), getDockerRunCommand(slave, cloud.getJenkinsUrl()), taskDefinition);
+        RunTaskResult runTaskResult = ecsService.runEcsTask(agent, template, cloud.getCluster(), getDockerRunCommand(agent, cloud.getJenkinsUrl()), taskDefinition);
 
         if (!runTaskResult.getFailures().isEmpty()) {
-            LOGGER.log(Level.WARNING, "[{0}]: Failure to run task with definition {1} on ECS cluster {2}", new Object[]{slave.getNodeName(), taskDefinition.getTaskDefinitionArn(), cloud.getCluster()});
+            LOGGER.log(Level.WARNING, "[{0}]: Failure to run task with definition {1} on ECS cluster {2}", new Object[]{agent.getNodeName(), taskDefinition.getTaskDefinitionArn(), cloud.getCluster()});
             for (Failure failure : runTaskResult.getFailures()) {
-                LOGGER.log(Level.WARNING, "[{0}]: Failure reason={1}, arn={2}", new Object[]{slave.getNodeName(), failure.getReason(), failure.getArn()});
+                LOGGER.log(Level.WARNING, "[{0}]: Failure reason={1}, arn={2}", new Object[]{agent.getNodeName(), failure.getReason(), failure.getArn()});
             }
-            throw new AbortException("Failed to run agent container " + slave.getNodeName());
+            throw new AbortException("Failed to run agent container " + agent.getNodeName());
         }
         Task task = runTaskResult.getTasks().get(0);
         String taskArn = task.getTaskArn();
 
-        LOGGER.log(Level.INFO, "[{0}]: Agent started with task arn : {1}", new Object[] { slave.getNodeName(), taskArn });
-        slave.setTaskArn(taskArn);
-        slave.setClusterArn(cloud.getCluster());
+        LOGGER.log(Level.INFO, "[{0}]: Agent started with task arn : {1}", new Object[] { agent.getNodeName(), taskArn });
+        agent.setTaskArn(taskArn);
+        agent.setClusterArn(cloud.getCluster());
 
         return task;
     }