+++ src/main/java/com/cloudbees/jenkins/plugins/amazonecs/ECSSlave.java
@@ -61,7 +61,7 @@ import hudson.slaves.ComputerLauncher;
 import hudson.slaves.OfflineCause;
 
 /**
- * This slave should only handle a single task and then be shutdown.
+ * This agent should only handle a single task and then be shutdown.
  *
  * @author <a href="mailto:nicolas.deloof@gmail.com">Nicolas De Loof</a>
  */
@@ -90,7 +90,7 @@ public class ECSSlave extends AbstractCloudSlave {
     private String taskArn;
 
     public ECSSlave(@Nonnull ECSCloud cloud, @Nonnull String name, ECSTaskTemplate template, @Nonnull ComputerLauncher launcher) throws Descriptor.FormException, IOException {
-        super(name, "ECS slave", template.getRemoteFSRoot(), 1, Mode.EXCLUSIVE, template.getLabel(), launcher, new OnceRetentionStrategy(5), Collections.EMPTY_LIST);
+        super(name, "ECS Agent", template.getRemoteFSRoot(), 1, Mode.EXCLUSIVE, template.getLabel(), launcher, new OnceRetentionStrategy(5), Collections.EMPTY_LIST);
         this.cloud = cloud;
         this.template = template;
     }